requirements.txt
# amqp==5.1.1
# asgiref==3.5.0
# async-timeout==4.0.2
# backports.zoneinfo==0.2.1
# billiard==3.6.4.0
# black==22.3.0
# celery==5.2.6
# certifi==2021.10.8
# cffi==1.15.0
# charset-normalizer==2.0.12
# click==8.1.2
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.2.0
# cryptography==36.0.2
# defusedxml==0.7.1
# Deprecated==1.2.13
# Django==4.0.3
# django-4-jet==1.0.8
# django-filter==21.1
# django-modeltranslation==0.17.7
# django-reverse-admin==2.9.6
# djangorestframework==3.13.1
# environs==9.5.0
# idna==3.3
# kombu==5.2.4
# marshmallow==3.15.0
# mypy-extensions==0.4.3
# numpy==1.22.3
# oauthlib==3.2.0
# overlap==0.1.0
# packaging==21.3
# pathspec==0.9.0
# Pillow==9.0.1
# platformdirs==2.5.1
# prompt-toolkit==3.0.29
# psycopg2-binary==2.9.3
# pycparser==2.21
# PyJWT==2.3.0
# pyparsing==3.0.9
# python-dotenv==0.20.0
# python3-openid==3.2.0
# pytz==2022.1
# redis==4.3.1
# redis-tools==0.0.2
# requests==2.27.1
# requests-oauthlib==1.3.1
# six==1.16.0
# social-auth-app-django==5.0.0
# social-auth-core==4.2.0
# sqlparse==0.4.2
# tomli==2.0.1
# typing-extensions==4.1.1
# urllib3==1.26.9
# vine==5.0.0
# wcwidth==0.2.5
# wrapt==1.14.1


amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
# backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.13
django-4-jet==1.0.8
django-celery-beat==2.2.1
django-filter==21.1
django-modeltranslation==0.17.7
django-reverse-admin==2.9.6
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
environs==9.5.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.15.0
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
overlap==0.1.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
# pkg_resources==0.0.0
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
redis==4.3.1
redis-tools==0.0.2
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1